move widget into subfolders
