+++ app/src/components/marker/marker.scss
@@ -25,8 +25,3 @@ p.price-marker {
         font-weight:700;
     }
 }
-
-.pickup-location{
-    width:400px;
-    margin: 20px;
-}
