+++ app/src/components/searchWidget/Autocomplete.tsx
@@ -3,6 +3,7 @@ import * as PropTypes from "prop-types";
 import classnames from "classnames";
 import { components } from "react-select";
 import { Async as AsyncSelect } from "react-select";
+import './widget.scss';
 
 // AsyncSelect custom components below
 const LoadingIndicator = () => <span className="loader alt" />;
