+++ app/src/components/searchWidget/widget.scss
@@ -1,0 +1,5 @@
+
+.pickup-location {
+    width: 400px;
+    margin: 20px;
+}
